Move blobs to different feature
