+++ frontend/src/features/experiences/skill-item.tsx
@@ -1,61 +1,11 @@
+import MotionBlob from "@features/blobs/motion-blob";
+import TextBlob from "@features/blobs/text-blob";
 import { Skill } from "@features/experiences/types/skill";
 import { motion } from "framer-motion";
-import { Children } from "react";
 
 type Props = {
 	skill: Skill;
 	vocationSkills: number[];
-};
-
-type Blob = {
-	text: string;
-	r: number;
-	cx: number;
-	cy: number;
-	path?: string;
-};
-
-const MotionBlob = ({ path, position, children }) => {
-	const transition = {
-		duration: 4,
-		repeat: Infinity,
-		repeatType: "reverse",
-		ease: "easeInOut",
-	};
-
-	let pathProps;
-	if (path) {
-		const offset = Math.random() * 20 + 10;
-		pathProps = {
-			style: {
-				offsetPath: `path("${path}")`,
-				offsetDistance: "var(--offset)",
-				offsetRotate: "0deg",
-			},
-			initial: { "--offset": `${position}%` },
-			animate: { "--offset": `${position + offset}%` },
-			transition: transition,
-		};
-	}
-	return <motion.g {...pathProps}>{children}</motion.g>;
-};
-
-const TextBlob = ({ cx = 0, cy = 0, r, text }: Blob) => {
-	return (
-		<>
-			<motion.circle cx={cx} cy={cy} r={r} />
-			<motion.text
-				x={cx}
-				y={cy}
-				textAnchor="middle"
-				dominantBaseline="middle"
-				fill="#fff"
-				overflow="hidden"
-				fontSize={r / 4}>
-				{text}
-			</motion.text>
-		</>
-	);
 };
 
 const BGCOLORS = [
@@ -73,7 +23,7 @@ const SkillItem = ({ skill, vocationSkills }: Props) => {
 	const x = commonFactor;
 	const y = commonFactor;
 	const radius = commonFactor / 2;
-	const offset = 10;
+	const offset = 20;
 	const path = `M ${x} ${y - radius - offset} a ${radius + offset} ${
 		radius + offset
 	} 0 1 1 0 ${(radius + offset) * 2} a ${radius + offset} ${
@@ -81,17 +31,17 @@ const SkillItem = ({ skill, vocationSkills }: Props) => {
 	} 0 1 1 0 ${-(radius + offset) * 2}`;
 
 	const surroundingBlobs = (subSkills: string[]) => {
-		return subSkills.map((skill, index) => {
-			const position = (index / subSkills.length) * 80;
+		return subSkills.map((subSkill, index) => {
+			const position = (index / subSkills.length) * 100;
 
 			return (
-				<MotionBlob key={index} path={path} position={position}>
+				<MotionBlob key={index} path={path} offsetDistance={position}>
 					<TextBlob
 						key={index}
 						cx={0}
 						cy={0}
 						r={radius / 2}
-						text={skill}
+						text={subSkill}
 					/>
 				</MotionBlob>
 			);
@@ -123,7 +73,7 @@ const SkillItem = ({ skill, vocationSkills }: Props) => {
 					</filter>
 				</defs>
 				<motion.g
-					className={`${bgColour} transition-colors`}
+					className={`${bgColour} transition-colors duration-500`}
 					filter="url(#goo)">
 					{skill.subskills && surroundingBlobs(skill.subskills)}
 					<TextBlob cx={x} cy={y} r={radius} text={skill.name} />
